+++ src/main/java/org/leviatanplatform/fractal/engine/calculators/Calculator.java
@@ -2,5 +2,5 @@ package org.leviatanplatform.fractal.engine.calculators;
 
 public interface Calculator {
 
-    double calculate(double real, double imaginary);
+    int calculate(double real, double imaginary);
 }
