+++ src/main/java/org/leviatanplatform/fractal/engine/calculators/JuliaCalculator.java
@@ -15,7 +15,7 @@ public class JuliaCalculator implements Calculator {
     }
 
     @Override
-    public double calculate(double real, double imaginary) {
+    public int calculate(double real, double imaginary) {
         return JuliaUtils.calculate(real, imaginary, c_real, c_imaginary, iterations);
     }
 }
